microgramm finish

+++ java_group_3_homework_55_Konstantin_Matushenko/src/main/java/com/example/hw50/controller/CommentController.java
@@ -1,8 +1,12 @@
 package com.example.hw50.controller;
 
 import com.example.hw50.model.Comment;
+import com.example.hw50.model.Publication;
+import com.example.hw50.model.User;
 import com.example.hw50.service.CommentServiceImpl;
+import com.example.hw50.service.PublicationServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,23 +17,29 @@ public class CommentController {
 
     @Autowired
     private CommentServiceImpl commentService;
+    @Autowired
+    private PublicationServiceImpl publicationService;
 
     @GetMapping
     public List<Comment> getAllComments() {
         return commentService.getAll();
     }
 
-    @PostMapping
+    @PostMapping(path = "/add")
     public Comment addComment(@RequestBody Comment comment) {
         return commentService.addComment(comment);
     }
 
-    @DeleteMapping("/{commentId}")
-    public Comment deleteComment(@PathVariable String commentId) {
-        if (commentService.existsById(commentId)) {
+    @DeleteMapping("del/{commentId}")
+    public Comment deleteComment(@PathVariable String commentId, Authentication authentication) {
+        if (commentService.existsById(commentId) && authentication != null) {
             Comment comment = commentService.getById(commentId);
-            commentService.deleteById(commentId);
-            return comment;
+            User user = (User) authentication.getPrincipal();
+            Publication publication = publicationService.getById(comment.getPublicationId());
+            if (user.getId().equals(publication.getUserId())) {
+                commentService.deleteById(commentId);
+                return comment;
+            }
         }
         return null;
     }
